Add files via upload
The document explains a logistic regression model developed to predict the occurrence of solar flares. The model has been programmed using Python. The data has been collected from www.spaceweatherlive.com which is open to the public. Firstly, the reasoning behind selecting this dataset and creating a model for the prediction of flares is given. The data is very uneven as solar flares are rare occasions which provides for a small number of data points for a positive event. Following that, the various steps and techniques for balancing the data such as Synthetic Minority Oversampling Technique (SMOTE), K-Means classification and One-Hot-Encoding are discussed. Finally, the accuracy of the model is reviewed and it is concluded that the logistic regression model is a good way to predict highly imbalanced datasets.